Gui: Added sound functionality
Note that Gui does not play sounds directly - instead it uses a callback when a sound is supposed to play (e.g. when a button is pressed).
Added data.close property to GuiButton constructor which makes the button close the closest closable when pressed.
Fixed GuiElement.setData being called GuiElement.getData. Derp...